Split parity smoke ownership gates
