+++ docs/CODE_QUALITY_FILE_GRADES.md
@@ -22,7 +22,7 @@ Generated by `scripts/grade-code-quality.py`. Grades are deterministic heuristic
 | `test:QuillCodeAppTests` | 214 | 33472 | 99 | A+ | `WorkspaceModelPickerSurfaceIntegrationTests.swift` A+<br>`WorkspaceReviewIntegrationTests.swift` A+<br>`WorkspaceBrowserIntegrationTests.swift` A+ |
 | `test:QuillCodeCoreTests` | 8 | 801 | 99 | A+ | `VerificationResultParserTests.swift` A+<br>`ModelCatalogCoreTests.swift` A+<br>`ToolSchemaCoreTests.swift` A+ |
 | `test:QuillCodeDesktopTests` | 5 | 1328 | 100 | A+ | `QuillCodeDesktopControllerSmokeTests.swift` A+<br>`QuillCodeApprovalNotificationTests.swift` A+<br>`QuillCodeDesktopRenderedSmokeTests.swift` A+ |
-| `test:QuillCodeParityTests` | 145 | 13754 | 100 | A+ | `ParitySwiftInteractionTargetAuditSupport.swift` A+<br>`ParityNativeInteractionContractGateTests.swift` A+<br>`ParitySmokeScriptGateTests.swift` A+ |
+| `test:QuillCodeParityTests` | 149 | 13763 | 100 | A+ | `ParitySwiftInteractionTargetAuditSupport.swift` A+<br>`ParityNativeInteractionContractGateTests.swift` A+<br>`ParityHTMLInteractionTargetAuditSupport.swift` A+ |
 | `test:QuillCodePersistenceTests` | 6 | 385 | 100 | A+ | `JSONProjectStoreTests.swift` A+<br>`ConfigStoreTests.swift` A+<br>`JSONAutomationStoreTests.swift` A+ |
 | `test:QuillCodeSafetyTests` | 5 | 1211 | 100 | A+ | `SafetyGitPullRequestPolicyTests.swift` A+<br>`SafetyDownloadPolicyTests.swift` A+<br>`SafetyPolicyTestCase.swift` A+ |
 | `test:QuillCodeToolsTests` | 32 | 4462 | 100 | A+ | `FileEditSessionGuardTests.swift` A+<br>`FileToolExecutorTests.swift` A+<br>`GitLocalToolExecutorTests.swift` A+ |
@@ -1080,6 +1080,7 @@ Generated by `scripts/grade-code-quality.py`. Grades are deterministic heuristic
 | A+ | 100 | 69 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityBrowserSnapshotGateTests.swift` | no automated issues |
 | A+ | 100 | 62 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityBrowserToolRendererGateTests.swift` | no automated issues |
 | A+ | 100 | 108 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityBrowserWorkflowGateTests.swift` | no automated issues |
+| A+ | 100 | 170 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityClickProbeFixtureSupport.swift` | no automated issues |
 | A+ | 100 | 70 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityCoreModelGateTests.swift` | no automated issues |
 | A+ | 99 | 73 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityDesktopBrowserAdapterGateTests.swift` | 1 lines >120 chars |
 | A+ | 99 | 270 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityDesktopControllerGateTests.swift` | 5 lines >120 chars |
@@ -1110,6 +1111,7 @@ Generated by `scripts/grade-code-quality.py`. Grades are deterministic heuristic
 | A+ | 100 | 148 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityInteractionTargetRegistryGateTests.swift` | no automated issues |
 | A+ | 100 | 15 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityInteractionTargetTempFileSupport.swift` | no automated issues |
 | A+ | 100 | 38 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityInteractionTargetTextSupport.swift` | no automated issues |
+| A+ | 99 | 195 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityLiveSmokeScriptGateTests.swift` | 1 lines >120 chars |
 | A+ | 100 | 89 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityMCPGateTests.swift` | no automated issues |
 | A+ | 99 | 110 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityMergeTrainGateTests.swift` | 1 lines >120 chars |
 | A+ | 99 | 73 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityModelGateTests.swift` | 2 lines >120 chars |
@@ -1124,7 +1126,9 @@ Generated by `scripts/grade-code-quality.py`. Grades are deterministic heuristic
 | A+ | 100 | 128 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityNativeSourceInteractionIconGestureTests.swift` | no automated issues |
 | A+ | 100 | 139 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityNativeSourceInteractionSpacingTests.swift` | no automated issues |
 | A+ | 100 | 48 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityNativeTopBarChromeGateTests.swift` | no automated issues |
+| A+ | 99 | 243 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityPackagedClickProbeSmokeGateTests.swift` | 5 lines >120 chars |
 | A+ | 99 | 67 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityPackagedMacOSSmokeGateTests.swift` | 1 lines >120 chars |
+| A+ | 99 | 276 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityPackagedWindowFixtureSupport.swift` | 1 lines >120 chars |
 | A+ | 100 | 36 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityPlaywrightAutomationGateTests.swift` | no automated issues |
 | A+ | 100 | 83 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityPlaywrightCommandPaletteGateTests.swift` | no automated issues |
 | A+ | 100 | 38 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityPlaywrightSettingsRuntimeGateTests.swift` | no automated issues |
@@ -1140,7 +1144,7 @@ Generated by `scripts/grade-code-quality.py`. Grades are deterministic heuristic
 | A+ | 100 | 69 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParitySlashSessionParserGateTests.swift` | no automated issues |
 | A+ | 100 | 55 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParitySlashThreadMemoryParserGateTests.swift` | no automated issues |
 | A+ | 100 | 89 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParitySlashWorkspaceParserGateTests.swift` | no automated issues |
-| A+ | 98 | 912 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParitySmokeScriptGateTests.swift` | 8 lines >120 chars |
+| A+ | 100 | 37 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParitySmokeScriptProcessSupport.swift` | no automated issues |
 | A+ | 97 | 635 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParitySwiftInteractionTargetAuditSupport.swift` | 9 lines >120 chars; 1 lines >160 chars |
 | A+ | 100 | 49 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityTerminalRendererGateTests.swift` | no automated issues |
 | A+ | 100 | 240 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityTestSupport.swift` | no automated issues |
